testing/mochitest/tests/test_cancelFullScreen.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 728893</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="start()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=728893">
  Mozilla Bug
</a>
<div id="div">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  /** Test for Bug 728893
      Checks if SimpleTest.finish cancels fullscreen mode
      The assertion happens after calling SimpleTest.finish
      Running the test without the harness won't work
  **/

  SpecialPowers.setBoolPref("full-screen-api.enabled", true);
  SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
                                 false);

  SimpleTest.waitForExplicitFinish();

  var div = document.getElementById("div");

  document.addEventListener("mozfullscreenchange", function (e) {
    if (document.mozFullScreenElement === div) {
      ok(true, "Calling SimpleTest.finish");
      SimpleTest.finish();
    }
    else {
      is(false, document.mozFullScreen,
        "SimpleTest.finish should cancel fullscreen mode");
    }
  }, false);

  function start() {
    SimpleTest.waitForFocus(function() {
      div.mozRequestFullScreen();
    });
  }

</script>
</pre>
</body>
</html>
